+++ sources/MotionMatching/Systems/ui_render.inl.cpp
@@ -4,7 +4,7 @@
 void briefing_ui_func();
 
 ecs::SystemDescription briefing_ui_descr("briefing_ui", {
-}, briefing_ui_func, ecs::SystemOrder::UI, (uint)(ecs::SystemTag::Game));
+}, briefing_ui_func, ecs::SystemOrder::UI, ecs::SystemTag::Game);
 
 void briefing_ui_func()
 {
@@ -22,7 +22,7 @@ ecs::SystemDescription motion_matching_statistic_descr("motion_matching_statisti
   {ecs::get_type_description<AnimationPlayer>("animationPlayer"), false},
   {ecs::get_type_description<Settings>("settings"), false},
   {ecs::get_type_description<ThirdPersonController>("thirdPersonController"), false}
-}, motion_matching_statistic_func, ecs::SystemOrder::UI, (uint)(ecs::SystemTag::Game));
+}, motion_matching_statistic_func, ecs::SystemOrder::UI, ecs::SystemTag::Game);
 
 void motion_matching_statistic_func()
 {
@@ -41,7 +41,7 @@ void current_anim_index_func();
 ecs::SystemDescription current_anim_index_descr("current_anim_index", {
   {ecs::get_type_description<AnimationPlayer>("animationPlayer"), false},
   {ecs::get_type_description<ThirdPersonController>("thirdPersonController"), false}
-}, current_anim_index_func, ecs::SystemOrder::UI, (uint)(ecs::SystemTag::Game));
+}, current_anim_index_func, ecs::SystemOrder::UI, ecs::SystemTag::Game);
 
 void current_anim_index_func()
 {
@@ -59,7 +59,7 @@ void menu_ui_func();
 ecs::SystemDescription menu_ui_descr("menu_ui", {
   {ecs::get_type_description<Settings>("settings"), false},
   {ecs::get_type_description<SettingsContainer>("settingsContainer"), false}
-}, menu_ui_func, ecs::SystemOrder::UIMENU, (uint)(ecs::SystemTag::Game));
+}, menu_ui_func, ecs::SystemOrder::UIMENU, ecs::SystemTag::Game);
 
 void menu_ui_func()
 {
@@ -77,7 +77,7 @@ void mm_early_text_perf_func();
 
 ecs::SystemDescription mm_early_text_perf_descr("mm_early_text_perf", {
   {ecs::get_type_description<Settings>("settings"), false}
-}, mm_early_text_perf_func, ecs::SystemOrder::UI, (uint)(ecs::SystemTag::Game));
+}, mm_early_text_perf_func, ecs::SystemOrder::UI, ecs::SystemTag::Game);
 
 void mm_early_text_perf_func()
 {
