+++ sources/common_systems/common/UI/common_ui.inl.cpp
@@ -5,7 +5,7 @@ void fps_ui_func();
 
 ecs::SystemDescription fps_ui_descr("fps_ui", {
   {ecs::get_type_description<EditorUI>("ui"), false}
-}, fps_ui_func, ecs::SystemOrder::UI, (uint)(ecs::SystemTag::Editor|ecs::SystemTag::Game));
+}, fps_ui_func, ecs::SystemOrder::UI, ecs::SystemTag::Editor|ecs::SystemTag::Game);
 
 void fps_ui_func()
 {
@@ -22,7 +22,7 @@ void debug_console_ui_func();
 
 ecs::SystemDescription debug_console_ui_descr("debug_console_ui", {
   {ecs::get_type_description<EditorUI>("ui"), false}
-}, debug_console_ui_func, ecs::SystemOrder::UI, (uint)(ecs::SystemTag::GameEditor|ecs::SystemTag::Debug));
+}, debug_console_ui_func, ecs::SystemOrder::UI, ecs::SystemTag::GameEditor|ecs::SystemTag::Debug);
 
 void debug_console_ui_func()
 {
