+++ sources/common_systems/common/UI/file_dialog.inl.cpp
@@ -5,7 +5,7 @@ void open_dialog_func();
 
 ecs::SystemDescription open_dialog_descr("open_dialog", {
   {ecs::get_type_description<EditorUI>("ui"), false}
-}, open_dialog_func, ecs::SystemOrder::UIMENU, (uint)( ecs::SystemTag::Editor));
+}, open_dialog_func, ecs::SystemOrder::UIMENU,  ecs::SystemTag::Editor);
 
 void open_dialog_func()
 {
